migration DB is create added profile

+++ src/main/java/com/neilo/springboot2/dao/Message.java
@@ -1,13 +1,12 @@
 package com.neilo.springboot2.dao;
 
-import javax.jws.soap.SOAPBinding;
 import javax.persistence.*;
 
 @Entity
 public class Message {
     @Id
     @GeneratedValue(strategy= GenerationType.AUTO)
-    private Integer id;
+    private Long id;
 
     private String text;
 
@@ -32,11 +31,11 @@ public class Message {
         this.author = user;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
